requirements.txt
pyqt5
PyQtWebEngine
folium
mysql-connector-python
python-dotenv